遊戯王公式DBのパック名から任意のパック名で登録できるように変更

+++ yugioh_cardDB/management/commands/GetCardDetail.py
@@ -2,6 +2,7 @@ from time import sleep
 import os
 import urllib3
 from tqdm import tqdm
+from bs4 import BeautifulSoup
 
 
 def getCardDetail():
@@ -15,13 +16,16 @@ def getCardDetail():
     while line:
         name, url = line.strip().split(",")
         cid = url[url.index("cid=") + 4:]
-        http = urllib3.PoolManager()
-        request = http.request('GET', "https://www.db.yugioh-card.com" + url + "&request_locale=ja")
-        html_file = open("yugioh_cardDB/texts/card_detail_html/" + cid + ".txt", "w", encoding="utf-8")
-        html_file.write(request.data.decode("utf-8"))
-        html_file.close()
+        if not os.path.isfile("yugioh_cardDB/texts/card_detail_html/" + cid + ".txt"):
+            http = urllib3.PoolManager()
+            request = http.request('GET', "https://www.db.yugioh-card.com" + url + "&request_locale=ja")
+            soup = BeautifulSoup(request.data,"html.parser")
+            html_data = soup.find("article")
+            html_file = open("yugioh_cardDB/texts/card_detail_html/" + cid + ".txt", "w", encoding="utf-8")
+            html_file.write(str(html_data))
+            html_file.close()
+            sleep(1)
         line = file.readline()
         bar.update(1)
-        sleep(1)
     bar.close()
     file.close()
